requirements.txt
absl-py==1.2.0
astor==0.8.1
astroid==1.6.6
autopep8==1.6.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2021.10.8
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
decorator==4.4.2
distlib==0.3.5
enum34==1.1.10
filelock==3.2.1
flask==2.1.3
funcsigs==1.0.2
gast==0.2.2
google-api-core==1.32.0
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.28.2
google-cloud-core==1.7.3
google-cloud-datastore==1.15.5
google-cloud-language==1.3.2
google-cloud-logging==1.15.3
google-cloud-spanner==1.19.3
google-cloud-storage==1.44.0
google-cloud-translate==2.0.4
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-pasta==0.2.0
google-resumable-media==1.3.3
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
h5py==2.10.0
httplib2==0.20.4
idna==2.10
importlib-metadata==2.1.3
importlib-resources==3.3.1
ipaddr==2.2.0
ipython==5.1.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
numpy
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==2.3.2
packaging==20.9
pandas==0.24.2
parso==0.7.1
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.0.2
pluggy==0.13.1
prompt-toolkit==1.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydocstyle==3.0.0
pyflakes==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rope==0.23.0
rsa==4.5
scandir==1.10.0
scipy==1.2.2
simplegeneric==0.8.1
singledispatch==3.7.0
six==1.16.0
snowballstemmer==2.2.0
tensorboard==2.9.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
termcolor==1.1.0
toml==0.10.2
traitlets==4.3.3
ujson
uritemplate==3.0.1
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yapf==0.32.0
zipp==1.2.0
